Moved intention test fixtures to the dedicated dir
