+++ src/test/fixtures/kphp_intentions/prettify-intention-3.fixture.php
