+++ src/test/fixtures/kphp_intentions/prettify-intention-5.qf.php
